+++ qiime2/plugins.py
@@ -490,16 +490,3 @@ class QIIMEArtifactAPIImporter:
 
 
 sys.meta_path += [QIIMEArtifactAPIImporter()]
-
-
-def assert_no_nans_in_tables(fh):
-    '''
-    Checks for NaNs present in any of the tables in the indicated file then
-    resets to the head of the file.
-    '''
-    from pandas import read_html
-
-    tables = read_html(fh)
-    for df in tables:
-        assert not df.isnull().values.any()
-    fh.seek(0)
